Added support for PID reuse in different programs, added -ifce switch for TS grab and fixed GCC warnings
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@5445 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ src/isomedia/box_code_base.c
@@ -333,7 +333,7 @@ GF_Err kind_Read(GF_Box *s,GF_BitStream *bs)
 			bytesToRead++;
 		}
 		ptr->schemeURI = gf_strdup(data);
-		schemeURIlen = strlen(data);
+		schemeURIlen = (u32) strlen(data);
 		if (bytesToRead > schemeURIlen+1) {
 			/* read the value */
 			char *data_value = data + schemeURIlen +1;
